Jetpack Connect: Separate flow types to a different module
