+++ client/jetpack-connect/authorize.js
@@ -37,7 +37,8 @@ import { authQueryPropTypes, getRoleFromScope } from './utils';
 import { decodeEntities } from 'lib/formatting';
 import { getCurrentUser } from 'state/current-user/selectors';
 import { isRequestingSite, isRequestingSites } from 'state/sites/selectors';
-import { JPC_PATH_PLANS, OFFER_RESET_FLOW_TYPES, REMOTE_PATH_AUTH } from './constants';
+import { JPC_PATH_PLANS, REMOTE_PATH_AUTH } from './constants';
+import { OFFER_RESET_FLOW_TYPES } from './flow-types';
 import { login } from 'lib/paths';
 import { recordTracksEvent as recordTracksEventAction } from 'state/analytics/actions';
 import { urlToSlug } from 'lib/url';
